added warning when Any time is returned from the function call and fixed return types

+++ mapie/metrics/regression.py
@@ -544,5 +544,5 @@ def regression_mwi_score(
     error_above: float = np.sum((y_true - y_pred_up)[y_true > y_pred_up])
     error_below: float = np.sum((y_pred_low - y_true)[y_true < y_pred_low])
     total_error = error_above + error_below
-    mwi = (width + total_error * 2 / (1 - confidence_level)) / len(y_true)
+    mwi = float((width + total_error * 2 / (1 - confidence_level)) / len(y_true))
     return mwi
